Testing Spring MVC Web Controllers - error page for delete student

+++ 7-Spring MVC Web Apps-MVC Controller Testing/3.00-starting-project/src/main/java/com/luv2code/springmvc/controller/GradebookController.java
@@ -39,6 +39,10 @@ public class GradebookController {
 
 	@GetMapping("/delete/student/{id}")
 	public String deleteStudent(@PathVariable int id, Model m){
+		if (!studentService.checkIfStudentIsNull(id)){
+			return "error";
+		}
+
 		studentService.deleteStudent(id);
 		Iterable<CollegeStudent> collegeStudents = studentService.getGradebook();
 		m.addAttribute("students", collegeStudents);
